src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
ascii_colors==0.7.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jwt==1.3.1
lightrag-hku==1.3.2
multidict==6.4.3
numpy==2.2.5
openai==1.75.0
packaging==25.0
pgvector==0.4.0
pipmaster==0.5.4
pluggy==1.5.0
prometheus_client==0.20.0
propcache==0.3.1
psutil==5.9.8
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
